app/requirements.txt
absl-py==1.4.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
chardet==4.0.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
idna==2.10
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.8.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-intel>=2.13.0,<2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.23.1
Werkzeug==2.3.6
wrapt==1.15.0
xgboost==1.3.3
zipp==3.16.2